refactor(analysis): extract near-dup fingerprint owner
